+++ src/Console/ConsoleProvider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Gacela\Console;
 
 use Gacela\Console\Domain\FilenameSanitizer\FilenameSanitizer;
+use Gacela\Console\Infrastructure\Command\CacheWarmCommand;
 use Gacela\Console\Infrastructure\Command\DebugContainerCommand;
 use Gacela\Console\Infrastructure\Command\ListModulesCommand;
 use Gacela\Console\Infrastructure\Command\MakeFileCommand;
@@ -36,6 +37,7 @@ final class ConsoleProvider extends AbstractProvider
             new MakeModuleCommand(),
             new ListModulesCommand(),
             new DebugContainerCommand(),
+            new CacheWarmCommand(),
         ]);
     }