+++ tests/Feature/Console/CacheWarm/TestModule/TestFacade.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\CacheWarm\TestModule;
+
+use Gacela\Framework\AbstractFacade;
+
+final class TestFacade extends AbstractFacade
+{
+    public function doSomething(): string
+    {
+        return 'test';
+    }
+}